+++ src/components/AboutTab.tsx
@@ -1,3 +1,6 @@
+/**
+ * The about tab in the manager.
+ */
 export default function AboutTab() {
   return <h1>This is the about tab</h1>;
 }
