+++ src/components/ErrorDisplay.tsx
@@ -1,11 +1,5 @@
 /**
- * The error display component.
- *
- * It is used for displaying error messages in a formatted block. It requires the
- * following props:
- *
- * - `title`: A title that will be displayed above the error message.
- * - `error`: The error message to display. Whitespace characters are preserved.
+ * The error display component for user widget errors.
  */
 export default function ErrorDisplay(props: { title: string; error: string }) {
   const { title, error } = props;
