+++ src/components/LogsTab.tsx
@@ -1,3 +1,6 @@
+/**
+ * The logs tab in the manager.
+ */
 export default function LogsTab() {
   return <h1>This is the logs tab</h1>;
 }
