+++ src/components/SettingsTab.tsx
@@ -1,3 +1,6 @@
+/**
+ * The global settings tab in the manager.
+ */
 export default function SettingsTab() {
   return <h1>This is the settings tab</h1>;
 }
